SPI Device Programming and Pulse Measurement
• Developed a driver for Ultrasonic Sensor connected to Intel Galileo
Development board, to trigger a measurement and read the pulse width
measure after trigger. This was made possible with the use of Interrupt
Handlers in Linux.
• Developed a driver for SPI and LED Display, where the messages were
transmitted over the SPI bus of Galileo to LED Display. Animations on
the display was controlled by the distance measured with the help of
sensor driver.